Update light cone utils
